Moved the asset modifier into standalone files
