+++ TooieRando/OptionData.h
@@ -1,4 +1,4 @@
-
+#pragma once
 #include <string>
 #include <vector>
 #include <algorithm>
